Route all assistant entry points through a shared openPositAssistantChat helper
The notebook panel, notebook cell quick-fix, and console Fix & Explain each
routed to posit-assistant.newChat with their own copy of the command id, payload
shape, and catch/log/toast block. Two of them also branched on
assistant.sidebarView and fell back to the built-in chat when the sidebar was
off -- logic that only exists because the deprecated positron-assistant
extension still ships. Since the buttons only render when posit-assistant is
installed (and newChat itself picks sidebar vs editor panel), that fallback is
unreachable as a working path and goes away with positron-assistant.

- Add positronAssistant/browser/positAssistantChat.ts owning POSIT_NEW_CHAT_COMMAND,
  the NewChatOptions/NewChatFile types, and openPositAssistantChat() (the shared
  try/catch/log/toast).
- Route all three consumers through it; drop the sidebarView/built-in-chat fallback.
- Move the command-routing unit test to the shared helper; drop the now-removed
  fallback tests from the console component test.
