+++ src/vs/workbench/contrib/positronNotebook/test/browser/askAssistantAction.vitest.ts
@@ -5,11 +5,7 @@
 
 /// <reference types="vitest/globals" />
 
-import { AskAssistantAction, openAssistantChat } from '../../browser/AskAssistantAction.js';
-import { ICommandService } from '../../../../../platform/commands/common/commands.js';
-import { INotificationService } from '../../../../../platform/notification/common/notification.js';
-import { NullLogService } from '../../../../../platform/log/common/log.js';
-import { stubInterface } from '../../../../../test/vitest/stubInterface.js';
+import { AskAssistantAction } from '../../browser/AskAssistantAction.js';
 
 /**
  * The Ask Assistant action is gated on the AI main switch (`config.ai.enabled`).
@@ -28,37 +24,5 @@ describe('AskAssistantAction', () => {
 	it('gates the editor toolbar button via the menu when clause', () => {
 		const menu = Array.isArray(desc.menu) ? desc.menu[0] : desc.menu;
 		expect(menu?.when?.serialize()).toContain('config.ai.enabled');
-	});
-});
-
-/**
- * The panel actions route to the standalone Posit Assistant via posit-assistant.newChat
- * (issue #14541). Pin the command id and payload so a regression to the inert built-in
- * chat, or a dropped payload field, fails loudly.
- */
-describe('openAssistantChat', () => {
-	it('routes the query to the Posit Assistant newChat command', async () => {
-		const executeCommand = vi.fn().mockResolvedValue(undefined);
-		const commandService = stubInterface<ICommandService>({ executeCommand });
-		const notificationService = stubInterface<INotificationService>({ error: vi.fn() });
-
-		await openAssistantChat(commandService, notificationService, new NullLogService(), 'Explain this notebook');
-
-		expect(executeCommand).toHaveBeenCalledWith('posit-assistant.newChat', {
-			prompt: 'Explain this notebook',
-			target: 'new',
-			behavior: 'submit',
-		});
-	});
-
-	it('notifies the user when the assistant command fails', async () => {
-		const executeCommand = vi.fn().mockRejectedValue(new Error('Posit Assistant is disabled.'));
-		const error = vi.fn();
-		const commandService = stubInterface<ICommandService>({ executeCommand });
-		const notificationService = stubInterface<INotificationService>({ error });
-
-		await openAssistantChat(commandService, notificationService, new NullLogService(), 'Explain this notebook');
-
-		expect(error).toHaveBeenCalledOnce();
 	});
 });
